Fixes a memory leak in qt6 pipeline blendMode switch and an image texture update issue
In case of a blend mode switch the pipeline was destroyed but not deleted, which caused a serious memory leak. This was changed so that we precreate standard blend mode pipelines and select the pipeline we like depending on the blend mode.
An additional issue was fixed. The issue was caused from a TextureTargetNode which was used as image or imageMesh target and beeing reused with a different texture. In that case we experienced a Direct3D Device Lost, / DEVICE HOLD /. The fixed reason was that the already allocated texture and buffers where not reseted, which subsequently could have caused in an buffer overflow and the write into unallocated graphic memory.