Avoid unloading Loop
At times, the application crashes when the user updates the Rive file due to excessive calls to updatePaintNode.

The crash occurs within the code responsible for updating the state machine:

The root cause appears to be multiple invocations of updatePaintNode() during the unloading phase,
leading to multiple emits of the loadFileAfterUnloading() signal.

The function loadFileAfterUnloading() triggers the state machine update code. During the unloading phase,
the artboard instance retains dangling pointers to the state machines, thereby causing the crash.